move Step class to lib/test-bot/step.rb
