+++ lib/test-bot.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require "test-bot/step"
